Fix warnings in non-unity static-analysis builds
Headers:
- Add `override` to defaulted destructors in ResponseMessage and 14
  WebSocket request/response subclasses (-Wsuggest-override).
- Cast size_t loop indices to int in LogParameters::operator== to match
  getLogPath's int parameter.

Sources:
- Combiner.cpp: fold `insertSuccess` local into `assert(pair.second)`
  directly; the local was unused under NDEBUG.
- JsonHelper.cpp / HttpClient.cpp / AwsSigV4Utility.cpp: static_cast<int>
  on size_t values passed to APIs that take int (send, HMAC_Init_ex,
  getLogPath).
- HttpClient.cpp: widen jsonStart/jsonEnd from int to size_t to match
  std::string::find return type.
- WebSocketppClientWrapper.cpp: mark unused OnMessage/OnTlsInit parameters
  as (void), drop an unused local errorMessage.
- GameLiftServerState.cpp: initialize m_onUpdateGameSession(nullptr) in
  both GAMELIFT_USE_STD and non-STD constructor init lists to silence
  uninitialized-member warnings.

+++ gamelift-server-sdk/source/aws/gamelift/internal/util/HttpClient.cpp
@@ -78,7 +78,7 @@ HttpResponse HttpClient::SendGetRequest(const std::string &url) {
 #endif
         }
 
-        ssize_t sent = send(sock, request.c_str(), request.length(), 0);
+        ssize_t sent = send(sock, request.c_str(), static_cast<int>(request.length()), 0);
         if (sent < 0) {
 #ifdef _WIN32
             throw std::runtime_error("Send failed, error number: " + std::to_string(WSAGetLastError()));
@@ -185,8 +185,8 @@ HttpResponse HttpClient::ParseHttpResponse(const std::string &response) {
     }
     if (response.find("Transfer-Encoding: chunked") != std::string::npos
         && response.find("Content-Type: application/json") != std::string::npos) {
-        int jsonStart = httpResponse.body.find("{");
-        int jsonEnd = httpResponse.body.find_last_of("}");
+        size_t jsonStart = httpResponse.body.find("{");
+        size_t jsonEnd = httpResponse.body.find_last_of("}");
         if (jsonStart != std::string::npos && jsonEnd != std::string::npos) {
             httpResponse.body = httpResponse.body.substr(jsonStart, jsonEnd - jsonStart + 2);
         }
